+++ enigma/cogs/basics.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from discord import Embed
 from discord.ext.commands import command, Cog, has_permissions, MissingPermissions, Context, cooldown, BucketType, \
-    CommandOnCooldown
+    CommandOnCooldown, CommandError, Command
 
 from enigma.settings import in_production
 from enigma.utils.colors import random_color
@@ -11,6 +11,93 @@ from enigma.utils.exceptions import NoError
 class Basics(Cog):
     def __init__(self, bot):
         self.bot = bot
+
+    # TODO - help command, before v0.3 due to multiple help embeds
+    @command(
+        name='help',
+        brief='Shows this message',
+        usage='[category|command]',
+        enabled=not in_production()
+    )
+    async def help(self, ctx: Context, arg: str = None):
+        help_em = Embed(
+            title=':grey_question: Help',
+            color=random_color()
+        )
+        help_category = 'You can also type `>help <category>` for more info on a category.'
+        help_command = 'Type `>help <command>` for more info on a command.'
+        if arg is None:
+            help_em.description = f'{help_category}\n{help_command}'
+            for cog in self.bot.cogs:
+                help_em.add_field(
+                    name=str(cog),
+                    value=', '.join(map(lambda x: f'`{x.name}`', self.bot.cogs[cog].get_commands())),
+                    inline=False
+                )
+        else:
+            cmd: Command = self.bot.get_command(arg)
+            if not cmd:
+                cog: Cog = self.bot.get_cog(arg.capitalize())
+                if not cog:
+                    help_em.description = f'No command or category called `{arg}` found.'
+                # Cog found
+                else:
+                    help_em.description = help_command
+                    help_em.add_field(
+                        name='Category',
+                        value=f'```\n{cog.qualified_name}\n```',
+                        inline=False
+                    )
+                    for cmd in cog.get_commands():
+                        try:
+                            if await cmd.can_run(ctx):
+                                help_em.add_field(
+                                    name=f'`{cmd}`',
+                                    value=cmd.brief or '\u200b'
+                                )
+                        except CommandError:
+                            pass
+            # Command found
+            else:
+                help_em.description = help_category
+                help_em.add_field(
+                    name='Command',
+                    value=f'```\n{cmd.name}\n```',
+                    inline=False
+                )
+                if cmd.description:
+                    help_em.add_field(
+                        name='Description',
+                        value=cmd.description,
+                        inline=False
+                    )
+                if cmd.aliases:
+                    help_em.add_field(
+                        name='Aliases',
+                        value=', '.join(map(lambda x: f'`{x}`', cmd.aliases)),
+                        inline=False
+                    )
+                if cmd.usage:
+                    help_em.add_field(
+                        name='Usage',
+                        value=f'`>{cmd.name} {cmd.usage}`',
+                        inline=False
+                    )
+                if cmd.help:
+                    help_em.add_field(
+                        name='Additional help',
+                        value=cmd.help,
+                        inline=False
+                    )
+                if cmd.aliases or cmd.usage:
+                    base = cmd.name if not cmd.aliases else f'[{"|".join([cmd.name, *cmd.aliases])}]'
+                    tail = cmd.usage or ''
+                    help_em.add_field(
+                        name='Classic representation',
+                        value=f'```\n>{base} {tail}\n```'
+                    )
+
+        await ctx.send(embed=help_em)
 
     @command(
         name='error',